requirements.txt
python-dotenv
google-cloud-secret-manager
google-api-core
google-genai
google-cloud-firestore
google-cloud-bigquery
firebase_admin
pydantic
demjson3
ruff
pytest
pydantic-settings
PyYAML
google-cloud-aiplatform
